REFACTOR-TS: only change the file extension `jsx` to `tsx`
