+++ components/Actor/ActorInfoCard.tsx
