+++ components/MovieDetail/MovieData.tsx
