+++ components/Trending/index.tsx
